refactor: 수정사항 반영 및 AI 패키지 구조 변경
