+++ auth/src/main/java/com/two_ddang/logistics/auth/infrastructure/config/UserFeignClientConfig.java
@@ -1,4 +1,4 @@
-package com.two_ddang.logistics.auth.config;
+package com.two_ddang.logistics.auth.infrastructure.config;
 
 import feign.Logger;
 import feign.RequestInterceptor;
