+++ auth/src/main/java/com/two_ddang/logistics/auth/infrastructure/exception/PasswordNotMatchException.java
@@ -1,4 +1,4 @@
-package com.two_ddang.logistics.auth.exception;
+package com.two_ddang.logistics.auth.infrastructure.exception;
 
 import com.two_ddang.logistics.core.exception.ApplicationException;
 import com.two_ddang.logistics.core.exception.ErrorCode;
