+++ auth/src/main/java/com/two_ddang/logistics/auth/user/UserFallback.java
@@ -1,7 +1,7 @@
 package com.two_ddang.logistics.auth.user;
 
 import com.two_ddang.logistics.auth.user.dto.SignUpRequestDto;
-import com.two_ddang.logistics.auth.user.dto.UserResponseDto;
+import com.two_ddang.logistics.auth.user.dto.UserRes;
 import feign.FeignException;
 import lombok.extern.slf4j.Slf4j;
 
@@ -15,7 +15,7 @@ public class UserFallback implements UserFeignClient {
     }
 
     @Override
-    public UserResponseDto getUserByUsername(String username) {
+    public UserRes getUserByUsername(String username) {
 
         if(cause instanceof FeignException.NotFound) {
             log.info("UserFallback 메소드에서 오류가 발생했습니다.");
