+++ auth/src/main/java/com/two_ddang/logistics/auth/user/UserFeignClient.java
@@ -1,8 +1,8 @@
 package com.two_ddang.logistics.auth.user;
 
-import com.two_ddang.logistics.auth.config.UserFeignClientConfig;
+import com.two_ddang.logistics.auth.infrastructure.config.UserFeignClientConfig;
 import com.two_ddang.logistics.auth.user.dto.SignUpRequestDto;
-import com.two_ddang.logistics.auth.user.dto.UserResponseDto;
+import com.two_ddang.logistics.auth.user.dto.UserRes;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,7 +14,7 @@ fallbackFactory = UserFallbackFactory.class)
 public interface UserFeignClient extends UserService {
 
     @GetMapping("/api/v1/users/{username}")
-    UserResponseDto getUserByUsername(@PathVariable("username") String username);
+    UserRes getUserByUsername(@PathVariable("username") String username);
 
     @PostMapping("/api/v1/users")
     void createUser(SignUpRequestDto requestDto);
