+++ auth/src/main/java/com/two_ddang/logistics/auth/user/UserService.java
@@ -1,11 +1,11 @@
 package com.two_ddang.logistics.auth.user;
 
 import com.two_ddang.logistics.auth.user.dto.SignUpRequestDto;
-import com.two_ddang.logistics.auth.user.dto.UserResponseDto;
+import com.two_ddang.logistics.auth.user.dto.UserRes;
 
 public interface UserService {
 
-    UserResponseDto getUserByUsername(String username);
+    UserRes getUserByUsername(String username);
 
     void createUser(SignUpRequestDto requestDto);
